requirements.txt
streamlit==1.38.0
google-generativeai==0.7.2
PyYAML==6.0
